- Moved xpath selector for global declraration - Removed import xpath selector in main test block - Updated path in container js to call for xpath file
