+++ example_casper/shared-file-test/end-to-end-one.js
@@ -1,5 +1,4 @@
 import { Helper } from './helper';
-import XPathSelector from './xpath-selector';
 
 fixture`End to End Test`;
